+++ Semana_4/node/src/modules/appointments/services/ListProvidersService.ts
@@ -1,7 +1,8 @@
-
 import IUsersRepository from "@users/repositories/IUsersRepository";
 import { inject, injectable } from "tsyringe";
+import ICacheProvider from "@shared/container/providers/CacheProvider/models/ICacheProvider";
 import Users from "@users/infra/Typeorm/entities/Users";
+import {classToClass} from 'class-transformer'
 
 interface IUserDTO {
   user_id: string;
@@ -11,17 +12,20 @@ interface IUserDTO {
 export default class ListProvidersService {
   constructor(
     @inject("UsersRepository")
-    private usersRepository: IUsersRepository
+    private usersRepository: IUsersRepository,
+    @inject("CacheProvider")
+    private cacheProvider: ICacheProvider
   ) {}
 
   public async execute({ user_id }: IUserDTO): Promise<Users[]> {
-    let users = await this.usersRepository.findAllProviders({
-      except_user_id: user_id,
-    });
+    let users = await this.cacheProvider.recover<Users[]>(`provider-list:${user_id}`);
 
-    users.forEach((user) => {
-      delete user.password;
-    });
-    return users;
+    if (!users)
+      users = await this.usersRepository.findAllProviders({
+        except_user_id: user_id,
+      });
+
+    await this.cacheProvider.save(`provider-list:${user_id}`, classToClass(users));
+    return classToClass(users);
   }
 }