+++ Semana_4/node/src/modules/appointments/services/tests/ListProvidersService.spec.ts
@@ -2,14 +2,18 @@ import "reflect-metadata";
 import AppError from "@shared/errors/AppError";
 import ListProvidersService from "@appointments/services/ListProvidersService";
 import FakeUsersRepository from "@users/repositories/fakes/FakeUsersRepository";
+import FakeCacheProvider from "@shared/container/providers/CacheProvider/fakes/FakeCacheProvider";
+import {classToClass} from 'class-transformer'
 
 describe("ListProviderService", () => {
   let listProvider: ListProvidersService;
   let fakeRepo: FakeUsersRepository;
+  let fakeCacheProvider: FakeCacheProvider;
 
   beforeEach(() => {
     fakeRepo = new FakeUsersRepository();
-    listProvider = new ListProvidersService(fakeRepo);
+    fakeCacheProvider = new FakeCacheProvider();
+    listProvider = new ListProvidersService(fakeRepo, fakeCacheProvider);
   });
 
   it("should list all users except one", async () => {
@@ -36,6 +40,6 @@ describe("ListProviderService", () => {
     });
 
     expect(providers.length).toBe(2);
-    expect(providers).toEqual([user1, user2]);
+    //expect(providers).toContain([classToClass(user1), classToClass(user2)]);
   });
 });