+++ Semana_4/node/src/modules/users/services/CreateUserService.ts
@@ -1,8 +1,9 @@
 import User from "../infra/Typeorm/entities/Users";
-import IHashProvider from '@users/providers/HashProvider/Models/IHashProvider'
+import IHashProvider from "@users/providers/HashProvider/Models/IHashProvider";
 import IUserRepository from "@users/repositories/IUsersRepository";
 import AppError from "../../../shared/errors/AppError";
 import { inject, injectable } from "tsyringe";
+import ICacheProvider from "@shared/container/providers/CacheProvider/models/ICacheProvider";
 
 interface IRequestDTO {
   name: string;
@@ -15,8 +16,10 @@ class CreateUserService {
   constructor(
     @inject("UsersRepository")
     private userRepository: IUserRepository,
-    @inject('HashProvider')
-    private hashProvider: IHashProvider
+    @inject("HashProvider")
+    private hashProvider: IHashProvider,
+    @inject("CacheProvider")
+    private cacheProvider: ICacheProvider
   ) {}
 
   public async execute({ name, email, password }: IRequestDTO): Promise<User> {
@@ -26,13 +29,15 @@ class CreateUserService {
 
     if (userExists) throw new AppError("User already used!");
 
-    let hashedPassword = await this.hashProvider.generateHash(password)
+    let hashedPassword = await this.hashProvider.generateHash(password);
 
     let user = await this.userRepository.create({
       name,
       email,
       password: hashedPassword,
     });
+
+    await this.cacheProvider.invalidatePrefix(`provider-list`)
 
     return user;
   }