+++ Semana_4/node/src/modules/users/services/tests/UpdateUserAvatarService.spec.ts
@@ -4,18 +4,30 @@ import UpdateUserAvatarService from "@users/services/UpdateUserAvatarService";
 import CreateUsersService from "@users/services/CreateUserService";
 import FakeStorageProvider from "@shared/container/providers/StorageProviders/fakes/FakeStorageProvider";
 import AppError from "@shared/errors/AppError";
+import FakeCacheProvider from "@shared/container/providers/CacheProvider/fakes/FakeCacheProvider";
 
 describe("UpdateUserAvatar", () => {
-  it("should create Insert user a new Avatar", async () => {
-    let fakerepository = new FakeUsersRepository();
-    let fakeHash = new FakeHashProvider();
-    let createService = new CreateUsersService(fakerepository, fakeHash);
-    let fakeStorage = new FakeStorageProvider();
-    let updateService = new UpdateUserAvatarService(
-      fakerepository,
-      fakeStorage
-    );
+  let fakerepository: FakeUsersRepository;
+  let fakeHash: FakeHashProvider;
+  let fakeCacheProvider: FakeCacheProvider;
+  let createService: CreateUsersService;
+  let fakeStorage: FakeStorageProvider;
+  let updateService: UpdateUserAvatarService;
 
+  beforeEach(() => {
+    fakerepository = new FakeUsersRepository();
+    fakeHash = new FakeHashProvider();
+    fakeCacheProvider = new FakeCacheProvider();
+    createService = new CreateUsersService(
+      fakerepository,
+      fakeHash,
+      fakeCacheProvider
+    );
+    fakeStorage = new FakeStorageProvider();
+    updateService = new UpdateUserAvatarService(fakerepository, fakeStorage);
+  });
+
+  it("should create Insert user a new Avatar", async () => {
     let user = await createService.execute({
       name: "Rafael",
       email: "rafael@email.com",
@@ -29,37 +41,23 @@ describe("UpdateUserAvatar", () => {
     });
 
     expect(user.avatar === avatar).toBeFalsy();
-    expect(user.avatar).toBe('ava');
+    expect(user.avatar).toBe("ava");
   });
 
   it("should not update avatar of a non authenticated user", async () => {
-    let fakerepository = new FakeUsersRepository();
-    let fakeStorage = new FakeStorageProvider();
-    let updateService = new UpdateUserAvatarService(
-      fakerepository,
-      fakeStorage
-    );
-
 
     expect(
-       updateService.execute({
+      updateService.execute({
         user_id: "",
         avatarFileName: "ava",
       })
     ).rejects.toBeInstanceOf(AppError);
   });
 
-  it("should delete avatar of user that already have an avatar",async() =>{
-    let fakerepository = new FakeUsersRepository();
-    let fakeHash = new FakeHashProvider();
-    let createService = new CreateUsersService(fakerepository, fakeHash);
-    let fakeStorage = new FakeStorageProvider();
-    let updateService = new UpdateUserAvatarService(
-      fakerepository,
-      fakeStorage
-    );
+  it("should delete avatar of user that already have an avatar", async () => {
 
-    let deleteFile =  jest.spyOn(fakeStorage,'deleteFile') // espiona para validar que foi chamado
+
+    let deleteFile = jest.spyOn(fakeStorage, "deleteFile"); // espiona para validar que foi chamado
 
     let user = await createService.execute({
       name: "Rafael",
@@ -67,7 +65,6 @@ describe("UpdateUserAvatar", () => {
       password: "123654",
     });
 
-   
     user = await updateService.execute({
       user_id: user.id,
       avatarFileName: "oldava",
@@ -82,7 +79,6 @@ describe("UpdateUserAvatar", () => {
 
     expect(user.avatar === oldAvatar).toBeFalsy();
 
-    expect(deleteFile).toHaveBeenCalledWith('oldava')
-
-  })
+    expect(deleteFile).toHaveBeenCalledWith("oldava");
+  });
 });